# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Prediction Game API",
    "description": "API for wallet-authenticated prediction gameplay, leaderboards, rounds, and predictions. Use Swagger UI to explore endpoints and test requests.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste a JWT like: Bearer <token>"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response returned by all API endpoints on failure.",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error class name (e.g. ValidationError, AuthenticationError, NotFoundError)",
            "example": "ValidationError"
          },
          "message": {
            "type": "string",
            "description": "Human-readable description of the error",
            "example": "walletAddress is required"
          },
          "code": {
            "type": "string",
            "description": "Machine-readable error code for programmatic handling",
            "example": "VALIDATION_ERROR"
          },
          "details": {
            "type": "array",
            "description": "Field-level validation details (present on validation errors only)",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "example": "walletAddress"
                },
                "message": {
                  "type": "string",
                  "example": "walletAddress is required"
                }
              },
              "required": [
                "field",
                "message"
              ]
            }
          }
        },
        "required": [
          "error",
          "message",
          "code"
        ]
      },
      "RateLimitResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        ],
        "example": {
          "error": "AppError",
          "message": "Too many requests from this IP, please try again after 15 minutes",
          "code": "RATE_LIMIT_EXCEEDED"
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/predictions/submit": {
      "post": {
        "tags": [
          "Predictions"
        ],
        "summary": "Submit a prediction",
        "description": "Submit a prediction for a round. Supports idempotency via Idempotency-Key header for safe retries.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "schema": {
              "type": "string"
            },
            "description": "Unique key for idempotent request handling (UUID recommended)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "roundId",
                  "amount",
                  "side"
                ],
                "properties": {
                  "roundId": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "number"
                  },
                  "side": {
                    "type": "string",
                    "enum": [
                      "up",
                      "down"
                    ]
                  },
                  "priceRange": {
                    "type": "object",
                    "properties": {
                      "min": {
                        "type": "number"
                      },
                      "max": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Prediction submitted"
          }
        }
      }
    },
    "/api/predictions/batch-submit": {
      "post": {
        "tags": [
          "Predictions"
        ],
        "summary": "Submit multiple predictions at once",
        "description": "Batch submit up to 50 predictions. Rate limit: **3 batch requests per minute per user** (stricter than single submit). On limit, responds with **429**.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "predictions"
                ],
                "properties": {
                  "predictions": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "roundId",
                        "amount"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Predictions processed"
          },
          "429": {
            "description": "Too many batch requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/predictions/user": {
      "get": {
        "tags": [
          "Predictions"
        ],
        "summary": "Get user predictions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of predictions"
          }
        }
      }
    },
    "/api/predictions/round/{roundId}": {
      "get": {
        "tags": [
          "Predictions"
        ],
        "summary": "Get predictions for a round",
        "parameters": [
          {
            "in": "path",
            "name": "roundId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of predictions"
          }
        }
      }
    },
...
  }
}